# Significant speedup of tool panel/homepage loading