Hi,
In short: here is a patch that makes the galaxy tool panel (and as a consequence the whole instance homepage) load much faster.

The whole story:
We had problems with 2 instances: loading the galaxy homepage could take around 10s before anything was displayed in the browser.
Applying all the recommendations in the wiki for performances, did not help very much.

I tried using the profiler available with the use_profile option in galaxy.ini, and found out that galaxy was taking a very long time preparing the list of tools and sections.

I looked at the code and found the problem: when loading the tool list, galaxy needs to associate to each tool the section name and id where it is placed. The way it was done until now, for each tool, it loops through the whole list of section and tools to find where it is placed. So galaxy end up spending a few seconds looping again and again in the tool list. And this is done everytime the tool panel needs to be sent to clients.

I wrote a patch that allows to cache every tools' position in the sections of the tool panel. The dict is created when starting galaxy and is updated whenever needed (tool removal/update).

I made some measures while loading the homepage on our instances (3 measures, the first one without browser cache):

> Without the patch:
> Total size: 2.1Mb/1.8Mb/1.8Mb
> Total time (including assets): 8.48s/7.83s/7.24s
> Base HTML only: 2.48s/2.54s/2.48s

> With the patch:
> Total size: 2.1Mb/1.8Mb/1.8Mb
> Total time (including assets): 6.16s/5.21s/5.46s total
> Base HTML only: 0.352s/0.327s/0.359s

=> for this instance, the homepage begins to appears 2 seconds earlier.
For another instance (with more tools), it's even better: the base HTML is loaded in ~800ms instead of 7.5s.

I can attach the profile output if someone is interested.

I applied the patch to 2 of our production instances and everything seems to work fine.

Tell me if you need any more info or if you see a possible problem with this patch.
Anthony