@jmchilton 2015-04-24
@abretaud Still digesting this - but it looks pretty cool though - thanks a bunch for tracking this down and for the fix!

Did you see there is a failing unit test in the Travis build?

    nosetests test/unit/tool_shed_unit_tests/test_tool_panel_manager.py

At first glance it seems like after deleting an entry from the ``_integrated_section_by_tool`` a new value needs to be generated if there were older tools in the lineage - any clue on the best way to do that? I can find some time to work on this - but if you have any ideas that would be great.