@abretaud 2015-04-24
Oh, I had forgotten to look at what happens in lib/tool_shed/galaxy_install/tools/tool_panel_manager.py...
I don't have the time to finish fixing it right now, but the tests are failing because we need to fill _integrated_section_by_tool in  lib/galaxy/tools/toolbox/base.py, around line 613. If you hardcode it like this, the tests should pass:

                self._tool_tag_manager.handle_tags( tool.id, elem )
                self.__add_tool( tool, load_panel_dict, panel_dict )
                self._integrated_section_by_tool[ guid ] = 'tid', 'test' # FIXME fill with real values
            # Always load the tool into the integrated_panel_dict, or it will not be included in the integrated_tool_panel.xml file.
            integrated_panel_dict.update_or_append( index, key, tool )

There may be also a bit of code to fix in the 2 remove functions.
If it can wait, I can fix this next week.
Anthony