@abretaud 2015-04-27
Hi!
Here's an updated patch: now it handles correctly the tool version removal, and it works with toolshed tools.
Anthony